requirements.txt
pandas
requests
python-dotenv
scikit-learn
lightgbm
pyyaml
tqdm
